+++ projects/fhi-angular-highcharts/src/lib/fhi-angular-highcharts.component.ts
@@ -6,6 +6,7 @@ import {
   HostListener,
   Input,
   OnChanges,
+  OnDestroy,
   Output,
 } from '@angular/core';
 
@@ -42,6 +43,7 @@ import { DiagramTypeGroup } from './models/diagram-type-group.model';
 import { FlaggedSerie } from './models/flagged-serie.model';
 import { DownloadService } from './services/download.service';
 import { MetadataForSeriesService } from './services/metadata-for-series.service';
+import { FhiDiagramRequirements } from './models/fhi-diagram-requirements.model';
 
 enum ControlsPopoverMenuActions {
   downloadSvg = 'downloadSvg',
@@ -53,10 +55,11 @@ enum ControlsPopoverMenuActions {
 @Component({
   selector: 'fhi-angular-highcharts',
   templateUrl: './fhi-angular-highcharts.component.html',
+  styleUrl: './fhi-angular-highcharts.component.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,
   standalone: false,
 })
-export class FhiAngularHighchartsComponent implements OnChanges {
+export class FhiAngularHighchartsComponent implements OnChanges, OnDestroy {
   private allSerieNames: string[] = [];
   private chartInstance!: Chart;
 
@@ -81,6 +84,7 @@ export class FhiAngularHighchartsComponent implements OnChanges {
 
   showDefaultChartTemplate: boolean;
   showDiagramTypeDisabledWarning: boolean;
+  diagramRequirements: FhiDiagramRequirements[] = [];
   showDiagramTypeNav: boolean;
   showDownloadButton: boolean;
   showTableOrientationButton: boolean;
@@ -133,6 +137,12 @@ export class FhiAngularHighchartsComponent implements OnChanges {
       this.updateDiagramState();
     } catch (error) {
       console.error(this.getErrorMsg(error));
+    }
+  }
+
+  ngOnDestroy() {
+    if (this.isFullscreen) {
+      document.body.style.overflow = '';
     }
   }
 
@@ -227,8 +237,6 @@ export class FhiAngularHighchartsComponent implements OnChanges {
     this.allSerieNames = [];
     this.flaggedSeries = [];
     this.metadataForSeriesService.resetMetadataForSeries();
-
-    this.isFullscreen = false;
   }
 
   private formatSerieName(name: string | Array<string>): string {
@@ -319,7 +327,12 @@ export class FhiAngularHighchartsComponent implements OnChanges {
       const msg = this.diagramTypeGroupService.getDiagramTypeDisabledWarningMsg(
         this.diagramOptionsInternal.activeDiagramType,
       );
-      console.warn(`Kan ikke vise diagramtype "${activeDiagramType}" fordi "${msg}"`);
+      this.diagramRequirements = this.diagramTypeGroupService.getDiagramRequirements(
+        this.activeDiagramTypeGroup.diagramType,
+      );
+      console.warn(
+        `Kan ikke vise diagramtype "${activeDiagramType}" fordi "${msg.message}" (Teknisk årsak: "${msg.warning}")`,
+      );
     }
   }