+++ src/MOCK_DB_DATA/library-items/library-items-shared-data.ts
@@ -170,6 +170,10 @@ export const LibraryItemsSharedData: LibraryItemsShared = {
     id: 'highcharts-all-inclusive',
     title: 'Highcharts all inclusive',
   },
+  HighchartsMapWithSlot: {
+    id: 'highcharts-map-with-slot',
+    title: 'Highcharts map with slot',
+  },
 
   Modal: {
     id: 'modal',