+++ src/app/views/shared/dynamic-library-examples/example-components/highcharts/mock-data.service.ts
@@ -18,6 +18,9 @@ import { DodsfallHjerteOgKarEtterKommune } from './mock-data/12.dodsfall-hjerte-
 
 // Data for testing/debugging while developing locally, do not show in dev or prod.
 import { TestData } from './mock-data/test-data-7';
+import { TestData11 } from './mock-data/test-data-11';
+import { TestData12 } from './mock-data/test-data-12';
+import { DodsfallHjerteOgKarEtterFylkeFlereAr } from './mock-data/11.dodsfall-hjerte-og-kar-etter-fylte-flere-ar';
 
 @Injectable({
   providedIn: 'root',
@@ -63,6 +66,15 @@ export class MockDataService {
         // Data for testing while developing locally, do not show in dev or prod.
         case MockData.TestData:
           dataSet.next(TestData);
+          break;
+        case MockData.TestData11:
+          dataSet.next(TestData11);
+          break;
+        case MockData.TestData12:
+          dataSet.next(TestData12);
+          break;
+        case MockData.DodsfallHjerteOgKarEtterFylkeFlereAr:
+          dataSet.next(DodsfallHjerteOgKarEtterFylkeFlereAr);
           break;
       }
     });