+++ src/app/views/shared/dynamic-library-examples/example-components/highcharts/mock-data/test-data-11.ts
@@ -1,0 +1,2 @@
+// Test data Empty
+export const TestData11 = [];